[FEAT] 현재 뉴스와 유사한 뉴스 조회 api 구현

+++ src/main/java/com/newstoss/news/adapter/in/web/dto/RelatedNewsDTO.java
@@ -15,5 +15,5 @@ public class RelatedNewsDTO {
     private String title;
     private String content;
     private String url;
-    private Long similarity;
+    private double similarity;
 }
